Add form component for  SOLID principle
